snowflake-data-loader/load_fhir_resource/requirements.txt
snowflake-snowpark-python
toml # for local development
pandas
requests
